Refactor config to .Options namespaces; update routing
Refactored all configuration-related classes to new .Options namespaces for logging, networking, and SDK modules. Updated all using directives accordingly. Removed redundant SuppressMessage attributes and related using statements from attribute classes. Moved PacketHandler to Internal.Compilation and PacketMetadata to Common.Networking.Packets. These changes improve modularity, clarity, and maintainability of the codebase.
