+++ src/Nalix.Logging/Internal/Console/ConsoleLoggerProvider.cs
@@ -15,7 +15,6 @@ using Nalix.Framework.Configuration;
 using Nalix.Framework.Injection;
 using Nalix.Framework.Options;
 using Nalix.Framework.Tasks;
-using Nalix.Logging.Configuration;
 using Nalix.Logging.Internal.Pooling;
 
 #if DEBUG
