+++ src/Nalix.Logging/Internal/File/FileLoggerProvider.cs
@@ -17,7 +17,6 @@ using Nalix.Framework.Configuration;
 using Nalix.Framework.Injection;
 using Nalix.Framework.Options;
 using Nalix.Framework.Tasks;
-using Nalix.Logging.Configuration;
 
 #if DEBUG
 [assembly: InternalsVisibleTo("Nalix.Logging.Tests")]
