+++ src/Nalix.Logging/NLogix.cs
@@ -8,7 +8,6 @@ using System.Runtime.CompilerServices;
 using System.Threading;
 using Microsoft.Extensions.Logging;
 using Nalix.Framework.Configuration;
-using Nalix.Logging.Configuration;
 using Nalix.Logging.Sinks;
 
 namespace Nalix.Logging;
