+++ src/Nalix.Logging/Sinks/BatchFileLogTarget.cs
@@ -5,7 +5,6 @@ using System;
 using System.Diagnostics;
 using System.Runtime.CompilerServices;
 using Microsoft.Extensions.Logging;
-using Nalix.Logging.Configuration;
 using Nalix.Logging.Formatters;
 using Nalix.Logging.Internal.File;
