+++ src/Nalix.Network/Internal/Limiters/ConnectionLimiter.cs
@@ -23,7 +23,6 @@ using Nalix.Framework.Memory.Pools;
 using Nalix.Framework.Options;
 using Nalix.Framework.Tasks;
 using Nalix.Framework.Time;
-using Nalix.Network.Configurations;
 using Nalix.Network.Connections;
 using Nalix.Network.Internal.Transport;
