+++ src/Nalix.Network/Internal/Time/TimingWheel.cs
@@ -18,7 +18,6 @@ using Nalix.Framework.Memory.Objects;
 using Nalix.Framework.Options;
 using Nalix.Framework.Tasks;
 using Nalix.Framework.Time;
-using Nalix.Network.Configurations;
 
 namespace Nalix.Network.Internal.Time;
