+++ src/Nalix.Network/Internal/Transport/AsyncCallback.cs
@@ -13,7 +13,6 @@ using Nalix.Common.Networking;
 using Nalix.Framework.Configuration;
 using Nalix.Framework.Injection;
 using Nalix.Framework.Memory.Objects;
-using Nalix.Network.Configurations;
 using Nalix.Network.Connections;
 using Nalix.Network.Internal.Pooling;
