+++ src/Nalix.SDK/Options/ITransportOptions.cs
@@ -1,9 +1,13 @@
 // Copyright (c) 2025-2026 PPN Corporation. All rights reserved.
 // Licensed under the Apache License, Version 2.0.
 
+
+// Copyright (c) 2025-2026 PPN Corporation. All rights reserved.
+// Licensed under the Apache License, Version 2.0.
+
 using Nalix.Common.Security;
 
-namespace Nalix.SDK.Configuration;
+namespace Nalix.SDK.Options;
 
 /// <summary>
 /// Defines the full set of options required to configure a transport connection.
