+++ src/Nalix.SDK/Transport/Internal/FRAME_READER.cs
@@ -16,7 +16,6 @@ using Nalix.Framework.Extensions;
 using Nalix.Framework.Injection;
 using Nalix.Framework.Memory.Buffers;
 using Nalix.Framework.Options;
-using Nalix.SDK.Configuration;
 
 namespace Nalix.SDK.Transport.Internal;
