+++ src/Nalix.SDK/Transport/UdpSession.cs
@@ -25,7 +25,6 @@ using Nalix.Framework.Options;
 using Nalix.Framework.Random;
 using Nalix.Framework.Security.Hashing;
 using Nalix.Framework.Time;
-using Nalix.SDK.Configuration;
 
 namespace Nalix.SDK.Transport;
